+++ serde_duper/tests/types_plain.rs
@@ -645,6 +645,58 @@ fn ipnet() {
 }
 
 #[test]
+#[cfg(feature = "jiff")]
+fn jiff() {
+    use jiff::{
+        Span, Timestamp, Zoned,
+        civil::{Date, DateTime, Time},
+    };
+    use serde_duper::types::jiff::{
+        DuperDate, DuperDateTime, DuperSpan, DuperTime, DuperTimestamp, DuperZoned,
+    };
+
+    #[derive(Debug, Serialize, Deserialize)]
+    struct Test {
+        #[serde(with = "DuperDate")]
+        date: Date,
+        #[serde(with = "DuperTime")]
+        time: Time,
+        #[serde(with = "DuperDateTime")]
+        dt: DateTime,
+        #[serde(with = "DuperTimestamp")]
+        ts: Timestamp,
+        #[serde(with = "DuperZoned")]
+        zoned: Zoned,
+        #[serde(with = "DuperSpan")]
+        span: Span,
+    }
+
+    let value = Test {
+        date: "1995-11-03".parse().unwrap(),
+        time: "01:33:00".parse().unwrap(),
+        dt: "2026-08-03T03:37:32".parse().unwrap(),
+        ts: "2024-06-19 15:22:45-04".parse().unwrap(),
+        zoned: "2024-08-10T23:14:00-04:00[America/New_York]"
+            .parse()
+            .unwrap(),
+        span: "P1MT1M".parse().unwrap(),
+    };
+    let serialized = serde_duper::to_string(&value).unwrap();
+    assert_eq!(
+        serialized,
+        r#"Test({date: PlainDate('1995-11-03'), time: PlainTime('01:33:00'), dt: PlainDateTime('2026-08-03T03:37:32'), ts: Instant('2024-06-19T19:22:45Z'), zoned: ZonedDateTime('2024-08-10T23:14:00-04:00[America/New_York]'), span: Duration('P1MT1M')})"#
+    );
+
+    let deserialized: Test = serde_duper::from_string(&serialized).unwrap();
+    assert_eq!(value.date, deserialized.date);
+    assert_eq!(value.time, deserialized.time);
+    assert_eq!(value.dt, deserialized.dt);
+    assert_eq!(value.ts, deserialized.ts);
+    assert_eq!(value.zoned, deserialized.zoned);
+    assert_eq!(value.span, deserialized.span.fieldwise());
+}
+
+#[test]
 #[cfg(feature = "regex")]
 fn regex() {
     use regex::{Regex, bytes::Regex as BytesRegex};